add dynamic rooms, redirect, refactor code
