+++ src/components/Sideboard/UserCard/Time/Time.scss
@@ -1,4 +1,4 @@
-@import "../../scss/variables";
+@import "../../../../scss/variables";
 
 .Time {
     color: $red;
