+++ src/components/Sideboard/UserCard/UserCard.js
@@ -1,9 +1,8 @@
 import React from 'react';
-import '../AvatarIcon/AvatarIcon';
-import '../Rooms/Rooms';
-import AvatarIcon from "../AvatarIcon/AvatarIcon";
-import Rooms from "../Rooms/Rooms";
-import Time from "../Time/Time";
+
+import AvatarIcon from '../../AvatarIcon/AvatarIcon';
+import Time from "./Time/Time";
+
 import './UserCard.scss';
 
 
@@ -11,12 +10,11 @@ const UserCard = () => {
     return (
             <div className="UserCard">
                 <div className="information-container">
-                    <Rooms/>
                     <Time/>
                 </div>
                 <div className="AvatarIcon-container">
                     <div className="inner-box">
-                    <AvatarIcon/>
+                        <AvatarIcon/>
                     </div>
                 </div>
             </div>
