+++ src/components/Sideboard/UserCard/UserCard.scss
@@ -1,4 +1,4 @@
-@import "../../scss/variables";
+@import "../../../scss/variables";
 
 .UserCard {
     height: 18vh;
